refactor: update queryClient import path and move queryClient definition to a new file
