+++ src/shared/libs/react-query/query-client.ts
